refactor(test): extract palette utils and add data-testid attributes
Prerequisites for palette/brand test coverage:
- Extract rgbToHex/hexToRgb from palette-manager.ts to
  src/utils/palette-utils.ts for unit testing without DOM
- Add 11 data-testid attributes to brand.astro and PalettePanel.astro
  per TEST_BEST_PRACTICES.md selector guidelines
