+++ src/components/brand/PalettePanel.astro
@@ -5,13 +5,13 @@ import { palettes } from '../../data/palettes';
 
 <SwatchControlStyles />
 
-<div class="palette-panel" id="palette-panel">
+<div class="palette-panel" id="palette-panel" data-testid="palette-panel">
   <div class="palette-panel__resize" id="panel-resize" title="Drag to resize"></div>
   <div class="palette-panel__edge">
-    <button class="palette-panel__toggle" id="panel-toggle" type="button" title="Toggle palette editor">
+    <button class="palette-panel__toggle" id="panel-toggle" type="button" title="Toggle palette editor" data-testid="palette-panel-toggle">
       <svg viewBox="0 0 64 64" fill="none" class="palette-panel__icon"><path d="M32 8 L56 56 L8 56 Z" fill="none" stroke="currentColor" stroke-width="5" stroke-linejoin="miter"/></svg>
     </button>
-    <div class="palette-panel__tabs" id="palette-tabs">
+    <div class="palette-panel__tabs" id="palette-tabs" data-testid="palette-tabs">
       {palettes.map((p, i) => (
         <button
           class={`palette-panel__tab${i === 0 ? ' palette-panel__tab--active' : ''}`}
@@ -23,17 +23,17 @@ import { palettes } from '../../data/palettes';
         </button>
       ))}
     </div>
-    <button class="palette-panel__popout" id="panel-popout-toggle" type="button" title="Pop out palette to all pages" aria-label="Toggle site-wide palette">
+    <button class="palette-panel__popout" id="panel-popout-toggle" type="button" title="Pop out palette to all pages" aria-label="Toggle site-wide palette" data-testid="palette-popout-toggle">
       <svg viewBox="0 0 64 64" fill="none" class="palette-panel__icon"><path d="M32 8 L56 56 L8 56 Z" fill="none" stroke="currentColor" stroke-width="5" stroke-linejoin="miter"/></svg>
     </button>
-    <button class="palette-panel__theme-toggle" id="panel-theme-toggle" type="button" title="Toggle dark theme" aria-label="Toggle dark theme">
+    <button class="palette-panel__theme-toggle" id="panel-theme-toggle" type="button" title="Toggle dark theme" aria-label="Toggle dark theme" data-testid="palette-theme-toggle">
       <svg viewBox="0 0 64 64" fill="none" class="palette-panel__icon"><path d="M32 8 L56 56 L8 56 Z" fill="none" stroke="currentColor" stroke-width="5" stroke-linejoin="miter"/></svg>
     </button>
   </div>
   <div class="palette-panel__body" id="panel-body">
     <div class="palette-panel__header">
       <span class="brutal-label" id="panel-palette-name">0. Current</span>
-      <button class="brutal-btn brutal-btn--secondary colors-reset-btn" type="button" id="reset-all">Reset</button>
+      <button class="brutal-btn brutal-btn--secondary colors-reset-btn" type="button" id="reset-all" data-testid="palette-reset-all">Reset</button>
     </div>
 
     <div class="palette-panel__section">
